chore: refactoring integration test for common use
