+++ packages/client-sdk-nodejs/test/integration/shared/topic-client.test.ts
@@ -1,0 +1,6 @@
+import {runTopicClientTests} from '@gomomento/common-integration-tests';
+import {SetupTopicIntegrationTest} from '../integration-setup';
+
+const {topicClient, Momento, IntegrationTestCacheName} =
+  SetupTopicIntegrationTest();
+runTopicClientTests(topicClient, Momento, IntegrationTestCacheName);
