+++ packages/common-integration-tests/src/index.ts
@@ -7,3 +7,4 @@ export * from './dictionary';
 export * from './set';
 export * from './sorted-set';
 export * from './item-get-type';
+export * from './topic-client';
