+++ packages/core/src/internal/clients/index.ts
@@ -1,3 +1,4 @@
 export * from './ping';
 export * from './cache';
 export * from './auth';
+export * from './pubsub';
