+++ packages/core/src/internal/clients/pubsub/index.ts
@@ -1,0 +1,2 @@
+export * from './ITopicClient';
+export * from './IPubsubClient';
